Make account balance to be updated right away after posting a transaction
